requirements.txt
Django==4.2.16
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==22.0.0
whitenoise==6.7.0
Pillow==10.4.0
reportlab==4.2.2
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
